elevator_ai_agent/requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
azure-cosmos>=4.5.0
requests>=2.31.0
pytz>=2023.3
pytest>=7.4.0
ruff>=0.0.291
black>=23.9.0